% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.matrix.sstapreg.R
\name{as.matrix.sstapreg}
\alias{as.matrix.sstapreg}
\title{Extract the posterior sample}
\usage{
\method{as.matrix}{sstapreg}(x, ..., pars = NULL)
}
\arguments{
\item{x}{sstaprego object}

\item{...}{Ignored.}

\item{pars}{vector of parameter names to subset from samples matrix}
}
\value{
A matrix  the dimensions of which depend on
  \code{pars} and \code{regex_pars}..
}
\description{
Provides samples of the posterior parameters in a
matrix data structure
}
\examples{
\donttest{
if (!exists("example_model")) example(example_model)
# Extract posterior sample after MCMC
draws <- as.matrix(example_model)
print(dim(draws))

# For example, we can see that the median of the draws for the intercept 
# is the same as the point estimate rstanarm uses
print(median(draws[, "(Intercept)"]))
print(example_model$coefficients[["(Intercept)"]])
}

}
\seealso{
\code{\link{sstapreg-methods}}
}
