% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bbnet_binomial_fit}
\alias{bbnet_binomial_fit}
\title{Bayesian Built Environment Network Model}
\usage{
bbnet_binomial_fit(y, X, rpc, max_q, prior_means, prior_scales, iter_max,
  max_treedepth, warm_up, seed)
}
\arguments{
\item{y}{vector of outcomes}

\item{X}{matrix of subject covariates}

\item{rpc}{array of regression parameter settings
rpc(0): the outcome distribution
rpc(1): boolean use_intercept value}

\item{max_q}{array of maximum number of distances per BEF}

\item{prior_means}{real valued vector for holding prior mean information
prior_means[0] = prior mean for intercept
prior_means[1] = prior mean for delta coefs ... and so on}

\item{prior_scales}{real valued vector for holding prior scale information
prior_scales[0] = prior scale for intercept
prior_scales[1] = prior scale for delta coefs ... and so on}

\item{iter_max}{maximum number of iterations}

\item{max_treedepth}{max treedepth for NUTS}

\item{warm_up}{number of iterations to use for epsilon adaptation}

\item{seed}{random number generator seed}

\item{num_basis}{array of number of basis functions per BEF}

\item{start_stop}{array of start-stop beta par indeces for D matrix}
}
\description{
Bayesian Built Environment Network Model
}
