% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbnetlm.R
\name{bbnet_lm}
\alias{bbnet_lm}
\title{Built Environment Networks Linear Model}
\usage{
bbnet_lm(formula, subject_data, subject_id = NULL, BEFs = NULL,
  basis_functions = NULL, dt_data = NULL, BEF_col_name = NULL,
  dt_col_name = NULL, ...)
}
\arguments{
\item{formula}{Similar as for \code{\link[stats]{lm}}.}

\item{subject_id}{string containing the common id column in both data and distance data and/or time_data}

\item{BEFs}{character vector containing BEF's coded in distance/time dataframe to include}

\item{...}{args for \code{\link[stats]{lm}}}

\item{data}{required data argument}

\item{distance_data}{distance dataframe containing up to four columns:
(1) subj_ID, (2) BEF_name and (3) Distance OR (4) Time between subj_ID and BEF}
}
\description{
Built Environment Networks Linear Model
}
