\name{bnet-package}
\alias{bnet-package}
\alias{bnet}
\docType{package}
\title{
\packageTitle{bnet}
}
\description{
\packageDescription{bnet}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{bnet}
\packageIndices{bnet}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{bnet}

Maintainer: \packageMaintainer{bnet}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
