% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bnet_lmer.R
\name{bnet_lmer}
\alias{bnet_lmer}
\title{Built Environment Networks Linear Mixed Effects Regression Model}
\usage{
bnet_lmer(formula, stap_formula, subject_data, subject_id = NULL,
  basis_functions = NULL, dt_data = NULL, BEF_col_name = NULL,
  distance_col_name = NULL, time_col_name = NULL, ...)
}
\arguments{
\item{formula}{Similar as for \code{\link[stats]{lm}}.}

\item{stap_formula}{See \code{\link[rstap]{stap_lm}}}

\item{subject_data}{required data argument containing subject level outcome and covariates}

\item{subject_id}{string name for the common id column in both data and distance data and/or time_data}

\item{dt_data}{distance dataframe containing up to four columns:
(1) subj_ID, (2) BEF_name and (3) Distance AND/OR (4) Time between subj_ID and BEF}

\item{BEF_col_name}{string name for the column containing the BEF labels in dt_data dataframe}

\item{distance_col_name}{string name for the column containing the subject-BEF distances in the dt_data dataframe}

\item{time_col_name}{string name for the the column containing the subject-BEF times in the dt_data dataframe}

\item{...}{args for \code{\link[stats]{lm}}}
}
\description{
Built Environment Networks Linear Mixed Effects Regression Model
}
