% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc-datasets.R
\docType{data}
\name{complex_longitudinal}
\alias{complex_longitudinal}
\title{Complex longitudinal benvo for use in \pkg{rsstap} examples and vignettes}
\format{
a benvo with 600 subjects, continuous outcome simulated using subject specific effect and temporal slope and between/within subject exposure effects from nearby "Healthy Food Stores"(HFS)
\describe{
\item{\code{subject_data}}{Includes pseudo Obesity outcome along with simulated exposure effects}
\item{ HFS data  }{ distances and times that subjects spend near HFS }
}
}
\usage{
complex_longitudinal
}
\description{
Complex longitudinal benvo for use in \pkg{rsstap} examples and vignettes
}
\seealso{
The data generation code on \href{github}{https://github.com/apeterson91/rsstap/tree/master/R/data-raw/complex_longitudinal.R}
}
\keyword{datasets}
