% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvcreg-methods.R
\name{family.rvcreg}
\alias{family.rvcreg}
\title{family method for rvcreg objects}
\usage{
\method{family}{rvcreg}(object, ...)
}
\arguments{
\item{object, ...}{See \code{\link[stats]{family}}}
}
\description{
family method for rvcreg objects
}
\keyword{internal}
