% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sstapreg-methods.R
\name{family.sstapreg}
\alias{family.sstapreg}
\title{family method for sstapreg objects}
\usage{
\method{family}{sstapreg}(object, ...)
}
\arguments{
\item{object, ...}{See \code{\link[stats]{family}}}
}
\description{
family method for sstapreg objects
}
\keyword{internal}
