% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sstapreg-methods.R
\name{formula.sstapreg}
\alias{formula.sstapreg}
\title{Formula method for sstapreg objects}
\usage{
\method{formula}{sstapreg}(x, ...)
}
\arguments{
\item{x}{a sstapreg object}

\item{...}{ignored currently}
}
\description{
Formula method for sstapreg objects
}
\keyword{internal}
