% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{get_grp_string}
\alias{get_grp_string}
\title{pastes string for regex expression}
\usage{
get_grp_string(BEFs)
}
\arguments{
\item{BEFs}{string of BEFs coding}
}
\description{
pastes string for regex expression
}
