% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvcspec.R
\name{get_rvcspec}
\alias{get_rvcspec}
\title{Retrieve STAP-RVC model specification from STAP model formula}
\usage{
get_rvcspec(mf, rf, benvo)
}
\arguments{
\item{mf}{formula for observation level mean estimation}

\item{rf}{formula for spatio-temporal level coefficient estimation}

\item{benvo}{Built Environment object - \code{\link[rbenvo]{benvo}} - containing data for model}
}
\value{
\code{\link{rvcspec}} object
}
\description{
Get rvcspec object which details the various components of the rvc model specification
}
