% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc-datasets.R
\docType{data}
\name{network_benvo}
\alias{network_benvo}
\title{Network Built Environment Feature Effects  dataset}
\format{
a benvo with 500 subjects, a pseudo BMI continuous outcome, pseudo continuous income and binary sex covariate measures with Fast Food Restaurant (FFR)
direct exposure effects and "indirect" network effects.
}
\usage{
network_benvo
}
\description{
Network Built Environment Feature Effects  dataset
}
\keyword{datasets}
