% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.rvcreg}
\alias{plot.rvcreg}
\title{Plotting a random varying  coefficients regression model}
\usage{
\method{plot}{rvcreg}(x, ...)
}
\arguments{
\item{x}{rvcreg object}
}
\description{
Plotting a random varying  coefficients regression model
}
