% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.sstapreg}
\alias{plot.sstapreg}
\title{Plot Spline Spatial Temporal Aggregated Predictors}
\usage{
\method{plot}{sstapreg}(x, stap_term = NULL, component = NULL)
}
\arguments{
\item{x}{sstapreg object}
}
\description{
Plot Spline Spatial Temporal Aggregated Predictors
}
