% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot3D}
\alias{plot3D}
\title{3D plots for rsstap models}
\usage{
plot3D(x, stap_term = NULL)
}
\arguments{
\item{x}{sstapreg object}

\item{stap_term}{string argument for which bef term to plot}
}
\description{
3D plots for rsstap models
}
\keyword{internal}
