% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot3D.sstapreg}
\alias{plot3D.sstapreg}
\title{3D plot for sstapreg objects}
\usage{
\method{plot3D}{sstapreg}(x, stap_term = NULL)
}
\arguments{
\item{x}{sstapreg object}

\item{stap_term}{string argument for which bef term to plot}
}
\description{
3D plot for sstapreg objects
}
