% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllMethods.R, R/plot.R
\name{plot_df_mer}
\alias{plot_df_mer}
\alias{plot_df}
\title{Parent Method for getting plot dataframe}
\usage{
plot_df_mer(object, pars = NULL)

plot_df(object, pars = NULL)
}
\description{
Parent Method for getting plot dataframe
}
