% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllMethods.R
\docType{methods}
\name{plot_df_mer,bbMod-method}
\alias{plot_df_mer,bbMod-method}
\title{Method for creating plot dataframe for bbMod objects}
\usage{
\S4method{plot_df_mer}{bbMod}(object, pars = NULL)
}
\description{
Method for creating plot dataframe for bbMod objects
}
