% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllMethods.R
\name{plot_df_mer}
\alias{plot_df_mer}
\title{sstapMod plot_df Method}
\usage{
plot_df_mer(object, pars = NULL)
}
\arguments{
\item{object}{sstapMod object}

\item{pars}{optional vector of parameter strings to subset the plotting dataframe}
}
\value{
plotting dataframe
}
\description{
sstapMod plot_df Method
}
