% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_directeffect.bbnet}
\alias{plot_directeffect.bbnet}
\title{Plots the Direct Spatio-Temporal Exposure of a bbnet Model}
\usage{
\method{plot_directeffect}{bbnet}(object, pars = NULL)
}
\arguments{
\item{object}{of type bbnet}

\item{pars}{optional parameter selection}
}
\description{
Plots the Direct Spatio-Temporal Exposure of a bbnet Model
}
