% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_effect.bbnet}
\alias{plot_effect.bbnet}
\title{Plotting Methods for bbnet objects}
\usage{
\method{plot_effect}{bbnet}(object, pars = NULL)
}
\arguments{
\item{object}{of type bbnet}

\item{pars}{optional parameter selection}
}
\description{
Plotting Methods for bbnet objects
}
