% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_effects}
\alias{plot_effects}
\alias{plot_effects.sstap}
\title{Plot Effects}
\usage{
plot_effects(object, pars = NULL)

\method{plot_effects}{sstap}(object, pars = NULL)
}
\arguments{
\item{object}{sstap object}

\item{pars}{optional vector of parameter strings to subset the plotting dataframe}
}
\description{
Plot Effects
}
\section{Methods (by class)}{
\itemize{
\item \code{sstap}: Plots the Direct Spatio-Temporal Exposure of a sstap Model
}}

