% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_effects.sstap}
\alias{plot_effects.sstap}
\title{Plots the Direct Spatio-Temporal Exposure of a sstap Model}
\usage{
\method{plot_effects}{sstap}(object, pars = NULL)
}
\arguments{
\item{object}{of type sstap}

\item{pars}{optional parameter selection}
}
\description{
Plots the Direct Spatio-Temporal Exposure of a sstap Model
}
