% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllMethods.R
\docType{methods}
\name{plot_effects_mer,bbMod-method}
\alias{plot_effects_mer,bbMod-method}
\title{Method for plotting bbMod objects}
\usage{
\S4method{plot_effects_mer}{bbMod}(object, pars = NULL)
}
\description{
Method for plotting bbMod objects
}
