% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllMethods.R
\docType{methods}
\name{plot_effects_mer,sstapMod-method}
\alias{plot_effects_mer,sstapMod-method}
\title{Method for plotting sstapMod objects}
\usage{
\S4method{plot_effects_mer}{sstapMod}(object, pars = NULL)
}
\arguments{
\item{object}{sstapMod object}

\item{pars}{optional vector of parameter strings to subset the plotting dataframe}
}
\value{
plotting dataframe
}
\description{
Method for plotting sstapMod objects
}
