% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllMethods.R
\name{plot_effects_mer}
\alias{plot_effects_mer}
\title{sstapMod Plot Effects Method}
\usage{
plot_effects_mer(object, pars = NULL)
}
\arguments{
\item{object}{sstapMod object}

\item{pars}{optional vector of parameter strings to subset the plotting dataframe}
}
\value{
plot with STAP effects
}
\description{
sstapMod Plot Effects Method
}
