% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotdf}
\alias{plotdf}
\alias{plotdf.sstapreg}
\title{Spatial Effect DataFrame}
\usage{
plotdf(x, stap_term = NULL, p = 0.95, grid = NULL)

\method{plotdf}{sstapreg}(x, stap_term = NULL, p = 0.95, grid = NULL)
}
\arguments{
\item{x}{sstapreg object}

\item{stap_term}{optional string for name of BEF smooth function to plot. 
Alternatively plots first BEF smooth function}

\item{p}{probability mass contained within uncertainty interval}

\item{grid}{by default is NULL corresponding to a .1 step grid from the min-max of the covariate space}
}
\description{
Returns a dataframe with the betas evaluated at either a default or given grid points
If no stap term is selected, the first stap component is plotted by default.
}
\section{Methods (by class)}{
\itemize{
\item \code{sstapreg}: 
}}

