% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc-datasets.R
\docType{data}
\name{poisson_benvo}
\alias{poisson_benvo}
\title{Small benvo for use in \pkg{rsstap} examples and vignettes.}
\format{
A benvo with 300 subjects, poisson outcome and nearby simulated "Fast Food Restaurants"(FFRs) and "Healthy Food Stores"(HFS)
\describe{
\item{\code{subject_data}}{includes Obese count, sex binary variable and continuous Income variable}
\item{\code{HFS data}}{simulated distances in (0,1) for "nearby HFS"}
\item{\code{FFR data}}{simulated distances in (0,1) for "nearby FFRs"}
}
}
\usage{
poisson_benvo
}
\description{
Small benvo for use in \pkg{rsstap} examples and vignettes.
}
\seealso{
the data generation code at <https://github.com/apeterson91/rsstap/tree/master/R/data-raw/poisson_benvo.R>
}
\keyword{datasets}
