% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_interval.R
\name{posterior_interval.sstapreg}
\alias{posterior_interval.sstapreg}
\alias{posterior_interval}
\title{Posterior uncertainty intervals}
\usage{
\method{posterior_interval}{sstapreg}(object, prob = 0.95, pars = NULL, ...)
}
\arguments{
\item{object}{sstapreg object}

\item{prob}{A number \eqn{p \in (0,1)}{p (0 < p < 1)} indicating the desired
probability mass to include in the intervals. The default is to report
\eqn{95}\% intervals (\code{prob=0.95}).}

\item{pars}{vector of parameter names}

\item{...}{ignored}
}
\value{
A matrix with two columns and as many rows as model parameters (or
  the subset of parameters specified by \code{pars}. 
  For a given value of \code{prob}, \eqn{p}, the columns
  correspond to the lower and upper \eqn{100p}\% interval limits and have the
  names \eqn{100\alpha/2}\% and \eqn{100(1 - \alpha/2)}\%, where \eqn{\alpha
  = 1-p}. For example, if \code{prob=0.9} is specified (a \eqn{90}\%
  interval), then the column names will be \code{"5\%"} and \code{"95\%"},
  respectively.
}
\description{
The \code{posterior_interval} function computes Bayesian posterior uncertainty
intervals. These intervals are often referred to as \emph{credible}
intervals. This text is inspired by the same function documentation from \pkg{rstanarm}.
}
\details{
\subsection{Interpretation}{
Unlike for a frenquentist confidence interval, it is valid to say that,
conditional on the data and model, we believe that with probability \eqn{p}
the value of a parameter is in its \eqn{100p}\% posterior interval. This
intuitive interpretation of Bayesian intervals is often erroneously applied
to frequentist confidence intervals. See Morey et al. (2015) for more details
on this issue and the advantages of using Bayesian posterior uncertainty
intervals (also known as credible intervals).
}
}
\references{
Gelman, A. and Carlin, J. (2014). Beyond power calculations:
assessing Type S (sign) and Type M (magnitude) errors. \emph{Perspectives on
Psychological Science}. 9(6), 641--51.

Morey, R. D., Hoekstra, R., Rouder, J., Lee, M. D., and Wagenmakers, E. 
(2016). The fallacy of placing confidence in confidence intervals. 
\emph{Psychonomic Bulletin & Review}. 23(1), 103--123.
}
