% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_predict.R
\name{posterior_predict.sstapreg}
\alias{posterior_predict.sstapreg}
\alias{posterior_predict}
\title{Posterior Predictive Distribution Estimate}
\usage{
\method{posterior_predict}{sstapreg}(object, type = "link", ...)
}
\arguments{
\item{object}{sstapreg object}

\item{type}{one of "link","response". See \code{\link[stats]{predict.glm}}.}

\item{...}{currently ignored}
}
\description{
Posterior Predictive Distribution Estimate
}
