% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{ppc}
\alias{ppc}
\alias{ppc.sstapreg}
\title{Posterior Predictive Checks}
\usage{
ppc(x, num_reps = 20)

\method{ppc}{sstapreg}(x, num_reps = 20)
}
\arguments{
\item{x}{a sstapreg object}

\item{num_reps}{number of yhat samples to plot}
}
\description{
Posterior Predictive Checks

Posterior Predictive Checks
}
\section{Methods (by class)}{
\itemize{
\item \code{sstapreg}: 
}}

\keyword{internal}
