% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print-and-summary.R
\name{print.rvcreg}
\alias{print.rvcreg}
\title{Print method for rvcreg objects}
\usage{
\method{print}{rvcreg}(x, digits = 2, ...)
}
\arguments{
\item{x}{sstapreg object}

\item{digits}{number of digits to round to}

\item{...}{ignored}
}
\description{
Print method for rvcreg objects
}
