% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print-and-summary.R
\name{print.sstapreg}
\alias{print.sstapreg}
\title{Print method for sstapreg objects}
\usage{
\method{print}{sstapreg}(x, digits = 1, ...)
}
\arguments{
\item{x}{sstapreg object}

\item{digits}{number of digits to round to}

\item{...}{ignored}
}
\description{
The \code{print} method for sstapreg objects displays a compact summary of the
fitted model.
}
