% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_interval.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{posterior_interval}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{rstantools}{\code{\link[rstantools]{posterior_interval}}}
}}

