% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc-datasets.R
\name{rsstap-datasets}
\alias{rsstap-datasets}
\alias{rsstap_demo}
\alias{rsstap_repeatedmeasures}
\title{Datasets for rstap examples}
\format{\describe{
\item{\code{rsstap_demo}}{
\itemize{
         \item \code{subject_data}
		\item -  \code{subj_id}: The subject unique identifier
		\item -  \code{sex}: Subject binary covariate
		\item -  \code{y}: Subject outcome
        \item \code{direct_distance_data}
		\item -  \code{subj_id}: Subject unique identifier
		\item -  \code{BEF}: BEF categorical coding
		\item -  \code{Distance}: Distance between subject and BEF
	\item Miscl
	\item -  alpha: simulated intercept
	\item -  delta: simulated binary covariate effect
	\item -  sigma: simulated standard deviation for residual normal errors
	\item -  true effect: simulated true effect
	}
}
Source: \href{https://biostatistics4socialimpact.github.io/rsstap/articles/Introduction}{Introductory Vignette}

\item{\code{rsstap_repeatedmeasures}}{
\itemize{
\item Still in Progress
}}
}}
\description{
Small datasets for use in \pkg{rstap} examples and vignettes.
}
