% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsstap-package.R
\docType{package}
\name{rsstap-package}
\alias{rsstap-package}
\alias{rsstap}
\title{The 'rsstap' package.}
\description{
Spline implimentation of Spatial Temporal Aggregated Predictors
}
\references{
Stan Development Team (2020). RStan: the R interface to Stan. R package version 2.19.3. https://mc-stan.org
}
