% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvc_lm.R
\name{rvc_lm}
\alias{rvc_lm}
\title{Randomly Varying Coefficients Spatial Temporal Aggregated Predictor Model}
\usage{
rvc_lm(mean_formula, rvc_formula, benvo, QR = TRUE, weights = NULL, ...)
}
\arguments{
\item{mean_formula}{formula for mean function}

\item{rvc_formula}{formula for random varying coefficients}

\item{benvo}{built environment object from the rbenvo package containing the relevant data}

\item{weights}{for unequal variances}

\item{...}{arguments for stan sampler}
}
\description{
Fits a spatial temporal aggregated predictor model where the 
spatial-temporal function is modeled using splines and
the function is modeled non additively as a function of proximate BEFs
}
