% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvcreg-methods.R
\name{rvcreg-methods}
\alias{rvcreg-methods}
\alias{nobs.rvcreg}
\alias{coef.rvcreg}
\alias{ses.rvcreg}
\alias{confint.rvcreg}
\alias{se.rvcreg}
\title{Methods for rvcreg objects}
\usage{
\method{nobs}{rvcreg}(x)

\method{coef}{rvcreg}(x)

ses.rvcreg(x)

\method{confint}{rvcreg}(object, ...)

\method{se}{rvcreg}(object, ...)
}
\arguments{
\item{object}{rvcreg object}

\item{...}{Ignored}
}
\description{
The methods documented on this page are actually some of the least important 
methods defined for \link{rvcreg} objects. The most 
important methods are documented separately, each with its own page. Links to
those pages are provided in the \strong{See Also} section, below.
This documentation is inspired by the \code{\link[rstanarm]{rstanarm}} package
}
\details{
The methods documented on this page are similar to the methods 
  defined for objects of class 'lm', 'glm', 'glmer', etc. However there are a
  few key differences:
  
\describe{
\item{\code{coef}}{
Medians are used for fixed effect point estimates. See the \emph{Point estimates} section
in \code{\link{print.sstapreg}} for more details.
}
\item{\code{se}}{
The \code{se} function returns fixed effects standard errors based on 
\code{\link{mad}}. See the \emph{Uncertainty estimates} section in
\code{\link{print.sstapreg}} for more details.
}
\item{\code{confint}}{
\code{confint} will throw an error because the
\code{\link{posterior_interval}} function should be used to compute Bayesian 
uncertainty intervals.
}}
}
