% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvcspec.R
\name{rvcspec}
\alias{rvcspec}
\title{Random Varying Coefficients Model Specification}
\usage{
rvcspec(
  stapless_formula,
  fake_formula,
  rf,
  components,
  terms,
  dimensions,
  between_withins,
  benvo
)
}
\arguments{
\item{stapless_formula}{from \code{\link{get_rvcspec}}}

\item{fake_formula}{list of ``fake'' formulas from \code{\link{get_sstapspec}}}

\item{rf}{formula for spatio-temporal level coefficient estimation}

\item{components}{Distance/Time dimension to be modeled}

\item{terms}{terms of BEFs}

\item{dimensions}{dimension of spline basis expansion}

\item{benvo}{built environment object - \code{\link[rbenvo]{benvo}} - containing data for model}

\item{between_within}{indicator variable determining whether to perform decomposition}
}
\description{
Random Varying Coefficients Model Specification
}
