% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sstapreg-methods.R
\name{se}
\alias{se}
\title{Extract standard errors}
\usage{
se(object, ...)
}
\arguments{
\item{object}{a Fitted sstapreg model object}

\item{...}{argument to method}
}
\value{
Standard errors of model parameters.
}
\description{
Generic function for extracting standard errors from fitted models.
}
\seealso{
\code{\link{se.sstapreg}}
}
\keyword{internal}
