% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sstap_df.R
\name{sstap_df}
\alias{sstap_df}
\title{Built Environment Network Direct Effect Matrix}
\usage{
sstap_df(stap_formula, subject_data, subject_id = NULL,
  basis_functions = NULL, dt_data = NULL, BEF_col_name = NULL,
  distance_col_name = NULL, time_col_name = NULL)
}
\arguments{
\item{stap_formula}{Similar as for \code{\link[rstap]{stap_lm}}.}

\item{subject_data}{required data argument}

\item{subject_id}{string containing the common id column in both data and distance data and/or time_data}

\item{basis_functions}{list with length equal to the number of BEFs in the stap_formula that specifies the basis function expansion}

\item{dt_data}{distance-time dataframe containing three columns (1) subj_ID, (2) BEF_name and (3) Distance between subj_ID and BEF}

\item{BEF_col_name}{string name of the column in the dt dataframe that contains the character vectors describing the BEFs}

\item{distance_col_name}{string name of the column in the dt dataframe that contains the distance values}

\item{time_col_name}{string name of the column in the dt dataframe that contains the time values}
}
\value{
matrix containing the aggregated distances/times across the basis functions provided for the respective BEFs
}
\description{
Built Environment Network Direct Effect Matrix
}
