% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sstap_glm.fit.R
\name{sstap_glm.fit}
\alias{sstap_glm.fit}
\title{Spline Spatial Temporal Aggregated Regression Linear Model Fit}
\usage{
sstap_glm.fit(y, Z, X, S, w = NULL, ...)
}
\arguments{
\item{y}{vector of outcomes}

\item{Z}{matrix of subject level covariates}

\item{X}{list of Smooth design matrices}

\item{S}{list of Smooth penalty/precision matrices}

\item{...}{arguments for stan sampler}
}
\description{
Spline Spatial Temporal Aggregated Regression Linear Model Fit
}
