% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sstap_glm.R
\name{sstap_glm}
\alias{sstap_glm}
\title{Spline Spatial Temporal Aggregated Predictor General Linear Model}
\usage{
sstap_glm(formula, benvo, weights = NULL, family = gaussian(), QR = TRUE, ...)
}
\arguments{
\item{formula}{Similar as for \code{\link[stats]{glm}} with the addition of \code{stap},\code{sap} \code{tap} terms as needed}

\item{benvo}{built environment \code{\link[rbenvo]{Benvo}} object from the \code{rbenvo} package containing the relevant subject-BEF data}

\item{weights}{vector of positive integer weights}

\item{family}{One of \code{\link[stats]{family}}  currently gaussian, binomial and poisson are implimented with identity, logistic and  log links currently.}

\item{QR}{boolean denoting whether or not to perform a QR decomposition on the design matrix.}

\item{...}{arguments for stan sampler}
}
\description{
Spline Spatial Temporal Aggregated Predictor General Linear Model
}
