% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sstap_lmer.R
\name{sstap_lmer}
\alias{sstap_lmer}
\title{Spline Spatial Temporal Aggregated Mixed Effects Regression Model}
\usage{
sstap_lmer(formula, benvo, ...)
}
\arguments{
\item{formula}{Similar as for \code{\link[lme4]{glmer}}.}

\item{benvo}{built environment object from the rbenvo package containing the relevant data}

\item{...}{optional arguments for stan sampler}
}
\description{
Spline Spatial Temporal Aggregated Mixed Effects Regression Model
}
