% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\docType{class}
\name{sstapMod-class}
\alias{sstapMod-class}
\alias{sstapMod}
\title{Lmer rsstap model objects}
\description{
A fitted lmer rsstap object. Inherits from \pkg{lme4}'s
\code{\link[lme4]{merMod}} so that most methods defined for this
parent class should work seamlessly within \pkg{rsstap} analysis
}
\section{Slots}{

\describe{
\item{\code{basis_functions}}{list of basis functions used for each STAP}

\item{\code{BEFs}}{A character vector containing the names of the BEFs included in the model fit}

\item{\code{stap_code}}{stap_code designation}

\item{\code{spaceranges}}{list of spatial range for each STAP}

\item{\code{timeranges}}{list of temporal ranges for each STAP}
}}

