% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\docType{class}
\name{sstapMod}
\alias{sstapMod}
\title{Lmer rsstap model objects}
\description{
A fitted lmer rsstap object. Inherits from \pkg{lme4}'s
\code{\link[lme4]{merMod}} so that most methods defined for this
parent class should work seamlessly within \pkg{rsstap} analysis
}
\section{Slots}{

\describe{
\item{\code{resp}}{An \code{lme4::\link[lme4]{lmResp}} object to store a
(mixed) model response variable}

\item{\code{stap_code}}{stap_code designation}

\item{\code{BEFs}}{A character vector containing the names of the BEFs included in the model fit}
}}

