% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sstapreg-methods.R
\name{sstapreg-methods}
\alias{sstapreg-methods}
\alias{VarCorr}
\alias{ngrps}
\alias{sigma}
\alias{nsamples}
\alias{coef.sstapreg}
\alias{confint.sstapreg}
\alias{fitted.sstapreg}
\alias{se.sstapreg}
\alias{nobs.sstapreg}
\alias{VarCorr.sstapreg}
\alias{ngrps.sstapreg}
\alias{vcov.sstapreg}
\alias{nsamples.sstapreg}
\title{Methods for sstapreg objects}
\usage{
\method{coef}{sstapreg}(object, ...)

\method{confint}{sstapreg}(object, ...)

\method{fitted}{sstapreg}(object, ...)

\method{se}{sstapreg}(object, ...)

\method{nobs}{sstapreg}(object, ...)

\method{VarCorr}{sstapreg}(x)

\method{ngrps}{sstapreg}(object, ...)

\method{vcov}{sstapreg}(object, correlation = FALSE, ...)

\method{nsamples}{sstapreg}(object, ...)
}
\arguments{
\item{object}{sstapreg object}

\item{...}{Ignored}

\item{x}{sstapreg object}

\item{correlation}{For \code{vcov}, if \code{FALSE} (the default) the
covariance matrix is returned. If \code{TRUE}, the correlation matrix is
returned instead.}
}
\description{
The methods documented on this page are actually some of the least important 
methods defined for \link[=sstapreg]{sstapreg} objects. The most 
important methods are documented separately, each with its own page. Links to
those pages are provided in the \strong{See Also} section, below.
}
\details{
The methods documented on this page are similar to the methods 
  defined for objects of class 'lm', 'glm', 'glmer', etc. However there are a
  few key differences:
  
\describe{
\item{\code{coef}}{
Medians are used for fixed effect point estimates. See the \emph{Point estimates} section
in \code{\link{print.sstapreg}} for more details.
}
\item{\code{se}}{
The \code{se} function returns fixed effects standard errors based on 
\code{\link{mad}}. See the \emph{Uncertainty estimates} section in
\code{\link{print.sstapreg}} for more details.
}
\item{\code{confint}}{
\code{confint} will throw an error because the
\code{\link{posterior_interval}} function should be used to compute Bayesian 
uncertainty intervals.
}}
}
