% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sstapspec-methods.R
\name{sstapspec-methods}
\alias{sstapspec-methods}
\alias{get_k}
\alias{get_terms_ix}
\alias{get_component}
\alias{has_bw}
\alias{get_smooth_obj}
\title{\code{sstapspec} Methods}
\usage{
get_k(x, term = NULL)

get_terms_ix(x, term = NULL)

get_component(x, term = NULL)

has_bw(x, term = NULL)

get_smooth_obj(x, term = NULL)
}
\arguments{
\item{x}{sstapspec object}

\item{term}{string of the BEF  used in the model fitting. 
Defaults to NULL which will retrieve the first stap term specified.}
}
\description{
\code{sstapspec} Methods
}
\details{
Methods for sstapspec objects. Function names are typically self-explanatory.
}
