% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sstapspec.R
\name{sstapspec}
\alias{sstapspec}
\title{Create STAP data structure}
\usage{
sstapspec(stapless_formula, fake_formula, stap_mat, benvo)
}
\arguments{
\item{stapless_formula}{from \code{\link{get_sstapspec}}}

\item{fake_formula}{list of ``fake'' formulas from \code{\link{get_sstapspec}}}

\item{stap_mat}{matrix of stap specification properties}

\item{benvo}{Built Environment object - \code{\link[rbenvo]{benvo}} - containing data for model}
}
\description{
Create STAP data structure
}
