% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stap.R
\name{stap}
\alias{stap}
\title{Create a stap object}
\usage{
stap(object)
}
\arguments{
\item{object}{a named list of objects containing information about the staps for a given model}
}
\value{
an object of class "stap"
}
\description{
Create a stap object
}
