% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print-and-summary.R
\name{summary.sstapreg}
\alias{summary.sstapreg}
\alias{print.summary.sstapreg}
\title{Summary method for sstapreg objects}
\usage{
\method{summary}{sstapreg}(object, probs = c(0.05, 0.5, 0.95), ..., digits = 1)

\method{print}{summary.sstapreg}(x, digits = max(1, attr(x, "print.digits")), ...)
}
\arguments{
\item{object}{sstapreg object}

\item{probs}{an optional numeric vector of probabilities pased to \code{\link[stats]{quantile}}}

\item{...}{ignored}

\item{digits}{Number of digits to use for formatting numbers when printing. 
When calling \code{summary}, the value of digits is stored as the 
\code{"print.digits"} attribute of the returned object.}

\item{x}{An object of class \code{"summary.sstapreg"}.}
}
\value{
The \code{summary} method returns an object of class 
  \code{"summary.sstapreg"}  which is a matrix of 
  summary statistics and diagnostics, with attributes storing information for use by the
  \code{print} method.
}
\description{
Summaries of parameter estimates and MCMC convergence diagnostics 
(Monte Carlo error, effective sample size, Rhat).
}
