% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sstapreg-methods.R
\name{waic}
\alias{waic}
\alias{waic.sstapreg}
\title{Calculate WAIC}
\usage{
waic(x)

\method{waic}{sstapreg}(x)
}
\arguments{
\item{x}{sstapreg object}
}
\description{
Calculate WAIC

Calculate WAIC
}
\section{Methods (by class)}{
\itemize{
\item \code{sstapreg}: calculate waic
}}

