% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbnetlm.R
\name{bbnet_lm}
\alias{bbnet_lm}
\title{Built Environment Networks Linear Model}
\usage{
bbnet_lm(formula, data, D, T_mat, iter = 2000, warm_up = 1000,
  chains = 1L, seed = NULL)
}
\arguments{
\item{formula}{Similar as for \code{\link[stats]{lm}}.}

\item{data}{same as for \code{\link[stats]{lm}}}

\item{D}{sparse basis distance matrix}

\item{iter}{number of iterations}

\item{warm_up}{number of warm up iterations}

\item{chains}{number of chains}

\item{seed}{random number generator}

\item{T}{translation matrix}
}
\description{
Built Environment Networks Linear Model
}
