% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.sstapreg}
\alias{plot.sstapreg}
\title{Plot Spline Spatial Temporal Aggregated Predictors}
\usage{
\method{plot}{sstapreg}(x, stap_term = NULL, p = 0.95, grid = NULL, ...)
}
\arguments{
\item{x}{sstapreg object}

\item{stap_term}{optional string for name of BEF smooth function to plot. 
Alternatively plots first BEF smooth function}

\item{p}{probability mass contained within uncertainty interval}

\item{grid}{by default is NULL corresponding to a .1 step grid from the min-max of the covariate space}

\item{...}{ignored}
}
\description{
Plots the smooth curve on a grid over the range of the stap smooth predictor
If no stap term is selected, the first stap component is plotted by default.
}
