% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.sstapreg}
\alias{plot.sstapreg}
\title{Plot Spline Spatial Temporal Aggregated Predictors}
\usage{
\method{plot}{sstapreg}(x, stap_term = NULL, component = NULL, ...)
}
\arguments{
\item{x}{sstapreg object}

\item{stap_term}{optional string for name of BEF smooth function to plot. 
Alternatively plots first BEF smooth function}

\item{component}{one of c("Distance","Time","Distance-Time") 
corresponding to the smooth function domain}

\item{...}{ignored}
}
\description{
Plot Spline Spatial Temporal Aggregated Predictors
}
