% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot3D}
\alias{plot3D}
\alias{plot3D.sstapreg}
\title{3D plots for rsstap models}
\usage{
plot3D(x, stap_term = NULL)

\method{plot3D}{sstapreg}(x, stap_term = NULL)
}
\arguments{
\item{x}{sstapreg object}

\item{stap_term}{string argument for which \code{stap()} bef term to plot}
}
\description{
3D plots for rsstap models
}
\section{Methods (by class)}{
\itemize{
\item \code{sstapreg}: 3d stap plot
}}

