% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_df}
\alias{plot_df}
\alias{plot_df.sstap}
\title{Plot Dataframe}
\usage{
plot_df(object, pars = NULL)

\method{plot_df}{sstap}(object, pars)
}
\arguments{
\item{object}{sstap object}

\item{pars}{optional vector of parameter strings to subset the plotting dataframe}
}
\description{
Plot Dataframe
}
\section{Methods (by class)}{
\itemize{
\item \code{sstap}: retrieves dataframe for plotting STAP effects
}}

