% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllMethods.R, R/plot.R
\name{plot_effects_mer}
\alias{plot_effects_mer}
\alias{plot_effects}
\title{Parent Method for plot_directeffect}
\usage{
plot_effects_mer(object, pars = NULL)

plot_effects(object, pars = NULL)
}
\description{
Parent Method for plot_directeffect
}
