% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print-and-summary.R
\name{print.sstapreg}
\alias{print.sstapreg}
\title{Print method for sstapreg objects}
\usage{
\method{print}{sstapreg}(x, digits = 1, ...)
}
\arguments{
\item{x}{sstapreg object}

\item{digits}{number of digits to round to}
}
\description{
The \code{print} method for stanreg objects displays a compact summary of the
fitted model. See the \strong{Details} section below for descriptions of the
different components of the printed output.
}
