% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sstap_glm.fit.R
\name{sstap_glm.fit}
\alias{sstap_glm.fit}
\title{Spline Spatial Temporal Aggregated Regression Generalized Linear Model Fit}
\usage{
sstap_glm.fit(y, Z, X, S, family, group = list(), QR = F, weights = NULL, ...)
}
\arguments{
\item{y}{vector/matrix of outcomes}

\item{Z}{matrix of subject level covariates}

\item{X}{list of Smooth design matrices}

\item{S}{list of Smooth penalty/precision matrices}

\item{family}{One of \code{\link[stats]{family}}  currently gaussian, binomial and poisson are implimented with identity, logistic and  log links currently.}

\item{group}{A list, possibly of length zero (the default), but otherwise
having the structure of that produced by \code{\link[lme4]{mkReTrms}} to
indicate the group-specific part of the model.}

\item{QR}{boolean denoting whether or not to perform a QR decomposition on the design matrix, note that this is an experimental feature and bugs are still being worked out.}

\item{weights}{for unequal variances}

\item{...}{arguments for stan sampler}
}
\description{
Spline Spatial Temporal Aggregated Regression Generalized Linear Model Fit
}
