% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sstap_glmer.R
\name{sstap_glmer}
\alias{sstap_glmer}
\title{Spline Spatial Temporal Aggregated Generalized Mixed Effects Regression Model}
\usage{
sstap_glmer(formula, benvo, family = gaussian(), QR = TRUE, ...)
}
\arguments{
\item{formula}{Similar as for \code{\link[lme4]{glmer}}.}

\item{benvo}{built environment object from the rbenvo package containing the relevant data}

\item{family}{One of \code{\link[stats]{family}}  currently gaussian, binomial and poisson are implimented with identity, logistic and  log links currently.}

\item{QR}{boolean denoting whether or not to perform a QR decomposition on the design matrix, note that this is an experimental feature and bugs are still being worked out.}

\item{...}{optional arguments for stan sampler}
}
\description{
Spline Spatial Temporal Aggregated Generalized Mixed Effects Regression Model
}
