% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sstap_lmer.R
\name{sstap_lmer}
\alias{sstap_lmer}
\title{Spline Spatial Temporal Aggregated Regression Model for Longitudinal Data}
\usage{
sstap_lmer(formula, stap_terms, benvo, weights = NULL, ...)
}
\arguments{
\item{formula}{Similar as for \code{\link[lme4]{glmer}}.}

\item{stap_terms}{list of BEF terms}

\item{benvo}{built environment object from the rbenvo package containing the relevant data}

\item{weights}{vector of positive integer weights}

\item{...}{arguments for stan sampler}
}
\description{
Spline Spatial Temporal Aggregated Regression Model for Longitudinal Data
}
